+++ lib/ruby_lokalise_api/collections/permission_templates.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module RubyLokaliseApi
+  module Collections
+    class PermissionTemplates < Base
+      ENDPOINT = RubyLokaliseApi::Endpoints::PermissionTemplatesEndpoint
+      RESOURCE = RubyLokaliseApi::Resources::PermissionTemplate
+      DATA_KEY = 'roles'
+    end
+  end
+end