+++ lib/ruby_lokalise_api/resources/permission_template.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module RubyLokaliseApi
+  module Resources
+    class PermissionTemplate < Base
+      MAIN_PARAMS = %i[nil].freeze
+      no_support_for %i[update destroy reload_data]
+    end
+  end
+end