+++ spec/support/expectations.rb
@@ -34,7 +34,8 @@ module Expectations
   def proper_endpoint(collection_obj)
     res, ep = first_collection_endpoint(collection_from(collection_obj))
 
-    params = res.class.const_get(:MAIN_PARAMS).to_array.map do |p|
+    main_params = res.class.const_get(:MAIN_PARAMS).to_array
+    params = main_params.map do |p|
       res.send(p).to_s
     end